Fix devices not appearing in games
Thanks to a message I got in discord apparently while SteamVR can tolerate some bad data, it seems that lots of games can't. I checked to see if the quat was normalised with OVR and it said it was but it was still invalid. I decided to try re-normalize it again and it seems like that fixed things so :)