Simplified vibration
This comes at the cost of weaker vibrations but a faster response time. I would try to figure out how the old code managed to increase the intensity of the vibrations but it is hard for me to read/understand. I may come back to this at another point in the future.